Conflicts:
	CS320_Lab03/derby.log

+++ CS320_Lab03/junit/edu/ycp/cs320/lab03/model/OrderTest.java
@@ -17,8 +17,8 @@ public class OrderTest {
 	 
 	private Restaurant r;
 	
-	private Float price1 = (float) 10.99;
-	private Float price2 = (float) 4.99;
+	private Double price1 = (Double) 10.99;
+	private Double price2 = (Double) 4.99;
 
 	private Menu testMenu1;
 	private String pizza = "pizza";
